Move all config related code into logic/config
