+++ packages/sanes-chrome-extension/src/logic/config/balances.ts
