+++ packages/sanes-chrome-extension/src/logic/config/chainsConfig/fetchConfig.ts
