+++ packages/sanes-chrome-extension/src/logic/config/chainsConfig/index.ts
