+++ packages/sanes-chrome-extension/src/logic/config/index.ts
@@ -1,0 +1,1 @@
+export * from './personafromconfig';
