+++ packages/sanes-chrome-extension/src/logic/config/wallet.ts
