+++ packages/sanes-chrome-extension/src/logic/persona/index.ts
@@ -1,2 +1,1 @@
-export * from './fromconfig';
 export * from './persona';
